+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/IntrospectionIT.java
@@ -54,7 +54,7 @@ public class IntrospectionIT {
 		__Schema schema = myQuery.__schema("{types {name fields(includeDeprecated:true) {name type {name}}}}");
 
 		// Verification
-		assertEquals(79, schema.getTypes().size());
+		assertEquals(82, schema.getTypes().size());
 		assertEquals("AllFieldCases", schema.getTypes().get(0).getName());
 		// As the order of fields seems to depend on the way this is compile, let's use a rough method to check the
 		// field names