+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateReservedJavaKeywordAsObjectField.java
@@ -1,0 +1,5 @@
+package org.allGraphQLCases.server.impl;
+
+public class DataFetchersDelegateReservedJavaKeywordAsObjectField {
+
+}