+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateReservedJavaKeywordAsObjectFieldImpl.java
@@ -1,0 +1,392 @@
+/**
+ * 
+ */
+package org.allGraphQLCases.server.impl;
+
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+
+import org.allGraphQLCases.server.Human;
+import org.allGraphQLCases.server.ReservedJavaKeywordAsObjectField;
+import org.allGraphQLCases.server.util.DataFetchersDelegateReservedJavaKeywordAsObjectField;
+import org.dataloader.DataLoader;
+import org.springframework.stereotype.Component;
+
+import graphql.schema.DataFetchingEnvironment;
+
+/**
+ * @author etienne-sf
+ *
+ */
+@Component
+public class DataFetchersDelegateReservedJavaKeywordAsObjectFieldImpl
+		implements DataFetchersDelegateReservedJavaKeywordAsObjectField {
+
+	@Override
+	public CompletableFuture<Human> _if(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin, String _new, String _package,
+			String _private) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _if(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin,
+			String _new, String _package, String _private) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _implements(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _implements(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _import(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _import(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _instanceof(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _instanceof(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _int(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _int(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _interface(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _interface(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _long(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _long(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _native(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _native(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _new(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _new(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _package(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _package(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _private(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _private(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _protected(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _protected(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _public(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _public(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _return(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _return(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _short(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _short(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _static(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _static(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _strictfp(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _strictfp(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _super(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _super(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _switch(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _switch(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _synchronized(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _synchronized(DataFetchingEnvironment dataFetchingEnvironment,
+			ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _this(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _this(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _throw(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _throw(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _throws(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _throws(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _transient(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _transient(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _try(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _try(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _void(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _void(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _volatile(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _volatile(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public CompletableFuture<Human> _while(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, Human> dataLoader, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public Human _while(DataFetchingEnvironment dataFetchingEnvironment, ReservedJavaKeywordAsObjectField origin) {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+}